fix: entity_sql stress test fixes (3 bugs)
Bug #1: PRAGMA journal_size_limit used Rust hex 0x4_000_000 — SQLite
doesn't understand underscored hex. Fixed to plain decimal 67108864.

Bug #2: SQL comments before SELECT caused misclassification — leading
-- and /* */ comments made classify() see empty first word, defaulting
to DML. Fixed: added strip_leading_comments() to classify(). Extracted
errors.rs from tools.rs for file-length compliance.

Bug #3: Multi-statement DML was non-atomic — first INSERT committed
even when second statement failed. Fixed: implicit BEGIN/COMMIT/ROLLBACK
wrapper for multi-statement batches, skipped when user starts with BEGIN.

40+ edge cases tested: CTEs, triggers, WITHOUT ROWID, RETURNING, ATTACH,
type coercion, transactions, VACUUM, UNION, subqueries, unicode, BLOBs.
