+++ .context-pilot/shared/entities/schema.sql
@@ -1,0 +1,90 @@
+-- Entity database dump (auto-generated)
+PRAGMA foreign_keys = OFF;
+
+CREATE TABLE IF NOT EXISTS _meta (
+            migration_id INTEGER PRIMARY KEY,
+            filename     TEXT NOT NULL,
+            applied_at   TEXT NOT NULL DEFAULT (datetime('now'))
+        );
+
+CREATE TABLE IF NOT EXISTS audit_log (id INTEGER PRIMARY KEY, action TEXT, table_name TEXT, row_id INTEGER, ts TEXT DEFAULT (datetime('now')));
+
+CREATE TABLE IF NOT EXISTS blobs_test (id INTEGER PRIMARY KEY, data BLOB);
+
+CREATE TABLE IF NOT EXISTS companies (
+    id INTEGER PRIMARY KEY,
+    name TEXT NOT NULL,
+    country TEXT,
+    founded INTEGER,
+    revenue REAL
+, ceo TEXT);
+
+CREATE TABLE IF NOT EXISTS people (
+    id INTEGER PRIMARY KEY,
+    name TEXT NOT NULL,
+    role TEXT,
+    company_id INTEGER REFERENCES companies(id)
+);
+
+CREATE TABLE IF NOT EXISTS tags (name TEXT PRIMARY KEY, color TEXT) WITHOUT ROWID;
+
+CREATE TABLE IF NOT EXISTS temp_test (id INTEGER PRIMARY KEY, val TEXT);
+
+CREATE TABLE IF NOT EXISTS us_companies_snapshot(
+  id INT,
+  name TEXT,
+  revenue REAL
+);
+
+INSERT OR IGNORE INTO "_meta" VALUES (1, '0001_20260604T161551.sql', '2026-06-04 16:15:51');
+INSERT OR IGNORE INTO "_meta" VALUES (2, '0002_20260604T161607.sql', '2026-06-04 16:16:07');
+INSERT OR IGNORE INTO "_meta" VALUES (3, '0003_20260604T161628.sql', '2026-06-04 16:16:28');
+INSERT OR IGNORE INTO "_meta" VALUES (4, '0004_20260604T161700.sql', '2026-06-04 16:17:00');
+INSERT OR IGNORE INTO "_meta" VALUES (5, '0005_20260604T161736.sql', '2026-06-04 16:17:36');
+INSERT OR IGNORE INTO "_meta" VALUES (6, '0006_20260604T161743.sql', '2026-06-04 16:17:43');
+INSERT OR IGNORE INTO "_meta" VALUES (7, '0007_20260604T161819.sql', '2026-06-04 16:18:19');
+INSERT OR IGNORE INTO "_meta" VALUES (8, '0008_20260604T161852.sql', '2026-06-04 16:18:52');
+INSERT OR IGNORE INTO "_meta" VALUES (9, '0009_20260604T161916.sql', '2026-06-04 16:19:16');
+INSERT OR IGNORE INTO "_meta" VALUES (10, '0010_20260604T161944.sql', '2026-06-04 16:19:44');
+INSERT OR IGNORE INTO "_meta" VALUES (11, '0011_20260604T161953.sql', '2026-06-04 16:19:53');
+INSERT OR IGNORE INTO "_meta" VALUES (12, '0012_20260604T162718.sql', '2026-06-04 16:27:18');
+INSERT OR IGNORE INTO "_meta" VALUES (13, '0013_20260604T162759.sql', '2026-06-04 16:27:59');
+
+INSERT OR IGNORE INTO "audit_log" VALUES (1, 'INSERT', 'companies', 8, '2026-06-04 16:17:47');
+INSERT OR IGNORE INTO "audit_log" VALUES (2, 'INSERT', 'companies', 9, '2026-06-04 16:18:14');
+INSERT OR IGNORE INTO "audit_log" VALUES (3, 'INSERT', 'companies', 10, '2026-06-04 16:24:52');
+INSERT OR IGNORE INTO "audit_log" VALUES (4, 'INSERT', 'companies', 11, '2026-06-04 16:26:19');
+INSERT OR IGNORE INTO "audit_log" VALUES (5, 'INSERT', 'companies', 12, '2026-06-04 16:28:42');
+
+INSERT OR IGNORE INTO "blobs_test" VALUES (1, X'cafe');
+
+INSERT OR IGNORE INTO "companies" VALUES (1, 'Acme Corp', 'France', 2019, 1650000.55, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (2, 'Globex Inc', 'US', 2015, 8200000, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (3, 'Initech', 'UK', 2021, 320000.75, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (4, 'Umbrella Corp', 'Japan', 1968, 99000000, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (5, 'Stark Industries', 'US', 2008, 45000000, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (7, 'O''Brien''s LLC', 'Ireland', NULL, NULL, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (8, 'Trigger Test Co', 'Germany', NULL, NULL, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (9, '日本電気株式会社', '日本', NULL, NULL, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (10, 'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA', 'X', NULL, NULL, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (11, 'Coercion Test', 'US', 'not-a-number', NULL, NULL);
+INSERT OR IGNORE INTO "companies" VALUES (12, 'Trigger+Return Co', 'Germany', 2024, NULL, NULL);
+
+INSERT OR IGNORE INTO "people" VALUES (1, 'Alice', 'Engineer', 1);
+INSERT OR IGNORE INTO "people" VALUES (2, 'Bob', 'Designer', 2);
+
+INSERT OR IGNORE INTO "tags" VALUES ('explicit_tx', '#333');
+INSERT OR IGNORE INTO "tags" VALUES ('low', '#00FF00');
+INSERT OR IGNORE INTO "tags" VALUES ('test_tx', '#111');
+INSERT OR IGNORE INTO "tags" VALUES ('urgent', '#FF00FF');
+INSERT OR IGNORE INTO "tags" VALUES ('wip', '#0000FF');
+
+INSERT OR IGNORE INTO "temp_test" VALUES (1, 'a');
+INSERT OR IGNORE INTO "temp_test" VALUES (2, 'b');
+INSERT OR IGNORE INTO "temp_test" VALUES (3, 'c');
+
+INSERT OR IGNORE INTO "us_companies_snapshot" VALUES (2, 'Globex Inc', 8200000);
+INSERT OR IGNORE INTO "us_companies_snapshot" VALUES (5, 'Stark Industries', 45000000);
+INSERT OR IGNORE INTO "us_companies_snapshot" VALUES (11, 'Coercion Test', NULL);
+
+PRAGMA foreign_keys = ON;
