+++ .context-pilot/shared/memories.yaml
@@ -34,14 +34,12 @@
   importance: high
   last_edited_ms: 0
 48f2140d3c2dee0e:
-  tl_dr: STOP before Phase 3 (Meilisearch sync + search scope) of entities implementation. User needs to review and edit the Meilisearch design doc BEFORE implementing the Meilisearch × Entities link. Go idle and ask user before starting any Phase 3 work.
-  importance: critical
+  tl_dr: 'Entities module COMPLETE — all 3 phases implemented: Phase 1 (scaffold), Phase 2 (DB/tool/panel/migrations), Phase 3 (Meilisearch sync + search scope + visualizer). Committed on ''entities'' branch and pushed. Design doc status: Implemented.'
+  importance: medium
   labels:
   - entities
-  - phase3
-  - blocker
-  - meilisearch
-  last_edited_ms: 1780585124407
+  - complete
+  last_edited_ms: 1780589593976
 8185ce57712b238d:
   tl_dr: 'ALWAYS push memories (memories.yaml) and tree descriptions (tree-descriptions.yaml) alongside code commits. When committing work, include a ''chore: sync memories and tree descriptions'' commit and push everything together. Never leave these unsync''d.'
   importance: critical
